Fix "pause crash"
The "pause crash" is a graphics crash commonly encountered in inaccurate emulators.
It does not happen on hardware.

In these inaccurate emulators, in areas where there is a minimap and it is visible (toggle with L), the game may randomly crash when pausing (by pressing start).
The crash is due to drawing the minimap, but the data used for it was replaced before the RSP got there.

The fix is to stop drawing the minimap earlier in the pausing process.